# Conflicts:
#	BlockDiagrams/Proteus.pdsprj
#	Code/Master/L1-Hardware-Project/L1-Hardware-Project.cproj
#	Code/Master/L1-Hardware-Project/src/defines.h
#	Code/Master/L1-Hardware-Project/src/sys_init.c
#	Code/Master/L1-Hardware-Project/src/sys_loop.c
#	Code/Master/L1-Hardware-Project/src/utils/I2C.c
#	Code/Master/L1-Hardware-Project/src/utils/SPI_Slave.c.orig
#	Code/Slave/L1-Hardware-Project/main.c
#	Code/Slave/L1-Hardware-Project/src/utils/GPS.c

+++ Code/Master/L1-Hardware-Project/src/utils/ultrasonic.c
@@ -8,7 +8,7 @@
 #include "../defines.h"
 
 
-int ultrazonic_distance(Pin trigPin, Pin echoPin, int timeout) {
+int ultrazonic_distance(Pin trigPin, Pin echoPin, unsigned long timeout) {
 	if (timeout == NULL) timeout = 500000;
 	pin_mode(trigPin, OUTPUT);
 	pin_mode(echoPin, INPUT_PULLUP);
